% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_python.R
\name{copernicus_configure_python}
\alias{copernicus_configure_python}
\title{Buscar e inicializar Python en cualquier sistema operativo (uso interno)}
\usage{
copernicus_configure_python()
}
\value{
La configuración de Python detectada (objeto py_config).
}
\description{
Busca el ejecutable de Python en rutas comunes y en el PATH del sistema.
Si lo encuentra, lo inicializa mediante reticulate.
}
\keyword{internal}
