% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_python.R
\name{copernicus_import_module}
\alias{copernicus_import_module}
\title{Import the copernicusmarine Python module (internal use)}
\usage{
copernicus_import_module(py)
}
\arguments{
\item{py}{Python configuration object.}
}
\value{
Imported copernicusmarine module.
}
\description{
Import the copernicusmarine Python module (internal use)
}
\keyword{internal}
