% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_download.R
\name{copernicus_is_ready}
\alias{copernicus_is_ready}
\title{Check if Copernicus Marine Python module is ready}
\usage{
copernicus_is_ready(verbose = TRUE)
}
\arguments{
\item{verbose}{Show detailed status information.}
}
\value{
TRUE if the Python module is loaded and ready.
}
\description{
Checks if the Python module is properly loaded to use Copernicus Marine.
Returns TRUE if the module exists.
}
