% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_copernicus.R
\name{copernicus_reinstall_package}
\alias{copernicus_reinstall_package}
\title{Reinstalar el paquete Python copernicusmarine}
\usage{
copernicus_reinstall_package()
}
\value{
Invisible TRUE si se instala correctamente.
}
\description{
Reinstala el módulo copernicusmarine en el entorno Python detectado.
}
