% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_download.R
\name{copernicus_test}
\alias{copernicus_test}
\title{Test Copernicus integration}
\usage{
copernicus_test(username = NULL, password = NULL)
}
\arguments{
\item{username}{Copernicus Marine username (optional).}

\item{password}{Copernicus Marine password (optional).}
}
\value{
TRUE if the test was successful.
}
\description{
Performs a small test download to validate that the whole system works.
}
