% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_read_dataframe.R
\name{copernicus_test_read_dataframe}
\alias{copernicus_test_read_dataframe}
\title{Probar funci칩n read_dataframe de Copernicus}
\usage{
copernicus_test_read_dataframe(username = NULL, password = NULL)
}
\arguments{
\item{username}{Usuario Copernicus Marine (opcional).}

\item{password}{Contrase침a Copernicus Marine (opcional).}
}
\value{
TRUE si la prueba fue exitosa.
}
\description{
Realiza una prueba peque침a de la funci칩n read_dataframe para validar que funcione.
}
