% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_copernicus.R
\name{setup_copernicus}
\alias{setup_copernicus}
\title{Set up Copernicus Marine integration (Python)}
\usage{
setup_copernicus(install_copernicus = TRUE)
}
\arguments{
\item{install_copernicus}{Logical. Should the copernicusmarine package be installed in Python if not available? Default: TRUE.}
}
\value{
Invisible TRUE if the configuration was successful.
}
\description{
Sets up the Python environment and loads the copernicusmarine module for use in R.
It is recommended to run this once per session before calling other functions in the package.
}
\examples{
setup_copernicus()
}
