% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_python.R
\name{copernicus_configure_python}
\alias{copernicus_configure_python}
\title{Search and initialize Python on any operating system (internal use)}
\usage{
copernicus_configure_python()
}
\value{
The detected Python configuration (py_config object).
}
\description{
Searches for the Python executable in common paths and the system PATH.
If found, it initializes it via reticulate.
}
\keyword{internal}
