% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_python.R
\name{copernicus_import_module}
\alias{copernicus_import_module}
\title{Importar el m贸dulo copernicusmarine de Python (uso interno)}
\usage{
copernicus_import_module(py)
}
\arguments{
\item{py}{Objeto de configuraci贸n Python.}
}
\value{
M贸dulo copernicusmarine importado.
}
\description{
Importar el m贸dulo copernicusmarine de Python (uso interno)
}
\keyword{internal}
