% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_python.R
\name{copernicus_install_package}
\alias{copernicus_install_package}
\title{Install the Python package copernicusmarine (internal use)}
\usage{
copernicus_install_package(py)
}
\arguments{
\item{py}{Python configuration object.}
}
\value{
Invisible TRUE if the installation is successful.
}
\description{
Install the Python package copernicusmarine (internal use)
}
\keyword{internal}
