% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_download.R
\name{copernicus_is_ready}
\alias{copernicus_is_ready}
\title{Verificar si Copernicus Marine est치 listo}
\usage{
copernicus_is_ready()
}
\value{
TRUE si el entorno est치 listo para descargar.
}
\description{
Retorna TRUE si el m칩dulo Python y archivo de configuraci칩n existen.
}
