% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_download.R
\name{copernicus_test}
\alias{copernicus_test}
\title{Probar integración Copernicus}
\usage{
copernicus_test(username = NULL, password = NULL)
}
\arguments{
\item{username}{Usuario Copernicus Marine (opcional).}

\item{password}{Contraseña Copernicus Marine (opcional).}
}
\value{
TRUE si la prueba fue exitosa.
}
\description{
Realiza una descarga de prueba pequeña para validar que todo el sistema funcione.
}
