% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_opendataset.R
\name{copernicus_test_open}
\alias{copernicus_test_open}
\title{Test opening of Copernicus dataset}
\usage{
copernicus_test_open(username = NULL, password = NULL)
}
\arguments{
\item{username}{Copernicus Marine username (optional, will try to get from stored credentials first).}

\item{password}{Copernicus Marine password (optional, will try to get from stored credentials first).}
}
\value{
TRUE if the test was successful.
}
\description{
Performs a test dataset opening to validate that the open_dataset function works.
Uses stored credentials if available.
}
