% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_opendataset.R
\name{copernicus_test_open}
\alias{copernicus_test_open}
\title{Probar apertura de dataset Copernicus}
\usage{
copernicus_test_open(username = NULL, password = NULL)
}
\arguments{
\item{username}{Usuario Copernicus Marine (opcional).}

\item{password}{Contraseña Copernicus Marine (opcional).}
}
\value{
TRUE si la prueba fue exitosa.
}
\description{
Realiza una prueba de apertura de dataset para validar que la función open_dataset funcione.
}
