% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.copernicus_env}
\alias{.copernicus_env}
\title{Copernicus private environment (internal use)}
\usage{
.copernicus_env()
}
\value{
A private environment in the global R space.
}
\description{
This function creates (if it doesn't already exist) and returns the internal environment
where the package stores persistent objects (e.g., the Python module, configurations, etc.).
It is used to avoid polluting the user's global environment and to maintain
references between the package's internal functions.
}
\keyword{internal}
