% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.copernicus_env}
\alias{.copernicus_env}
\title{Ambiente privado de Copernicus (uso interno)}
\usage{
.copernicus_env()
}
\value{
Un environment privado en el espacio global de R.
}
\description{
Esta función crea (si no existe) y retorna el ambiente interno donde el paquete
guarda objetos persistentes (por ejemplo, el módulo Python, configuraciones, etc.).
Se utiliza para evitar contaminación en el espacio global del usuario y mantener
las referencias entre funciones internas del paquete.
}
\keyword{internal}
