% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_copernicus.R
\name{setup_copernicus}
\alias{setup_copernicus}
\title{Configura la integración con Copernicus Marine (Python)}
\usage{
setup_copernicus(install_copernicus = TRUE)
}
\arguments{
\item{install_copernicus}{Lógico. ¿Instalar el paquete copernicusmarine en Python si no está disponible? Default: TRUE.}
}
\value{
Invisible TRUE si la configuración fue exitosa.
}
\description{
Configura el entorno Python y carga el módulo copernicusmarine para su uso en R.
Se recomienda ejecutar una sola vez por sesión antes de llamar a otras funciones del paquete.
}
\examples{
setup_copernicus()
}
